pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "repoinsight-mcp"
version = "1.0.0"
description = "Local-first MCP server for GitHub repository analysis"
requires-python = ">=3.10"
dependencies = [
    "requests>=2.31.0",
    "PyGithub>=2.1.1",
    "gitpython>=3.1.40",
    "fastapi>=0.104.1",
    "uvicorn>=0.24.0",
    "pydantic>=2.5.0",
    "sse-starlette>=1.8.2",
    "chardet>=5.2.0",
    "pygments>=2.17.0",
    "orjson>=3.9.10",
    "rich>=13.5.2,<14.0.0",
    "tqdm>=4.66.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.3",
    "black>=23.12.0",
    "mypy>=1.7.1",
]

[project.scripts]
repoinsight-mcp = "repoinsight_mcp.main:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.black]
line-length = 100
target-version = ["py310"]

[tool.mypy]
python_version = "3.10"
strict = true
warn_return_any = true
warn_unused_configs = true
